% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{read_encrypted_credentials}
\alias{read_encrypted_credentials}
\title{Read the encrypted credentials file}
\usage{
read_encrypted_credentials(credentials_filename = "credentials.yml.enc")
}
\arguments{
\item{credentials_filename}{the optional filename of the encrypted credentials file.}
}
\description{
Read the encrypted credentials file
}
