% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{use_encrypted_credentials}
\alias{use_encrypted_credentials}
\title{Setup encrypted credentials}
\usage{
use_encrypted_credentials()
}
\description{
This creates an unencrypted master.key file and an empty encrypted
credentials file.
}
