% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{write_encrypted_credentials}
\alias{write_encrypted_credentials}
\title{Update the encrypted credentials file}
\usage{
write_encrypted_credentials(object,
  credentials_filename = "credentials.yml.enc")
}
\arguments{
\item{object}{an object that is stored encrypted on disk. The object will be
converted to yaml using the yaml package. Please make sure the content can be
properly represented as yaml.}

\item{credentials_filename}{the optional filename of the encrypted credentials file.}
}
\description{
Update the encrypted credentials file
}
